# src/Ai.Tlbx.MidTerm/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "MidTerm API",
    "description": "MidTerm terminal multiplexer API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/change-password": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sessions": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionListDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/CreateSessionRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionInfoDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}": {
      "delete": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sessions/{id}/resize": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResizeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/sessions/{id}/name": {
      "put": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "auto",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/sessions/{id}/upload": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/files/register": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileRegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/files/check": {
      "post": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FileCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileCheckResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/list": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DirectoryListResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/view": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/files/download": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/files/resolve": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deep",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResolveResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/logs/read": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "file",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lines",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "fromEnd",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogReadResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/logs/tail": {
      "get": {
        "tags": [
          "Ai.Tlbx.MidTerm.OpenApi"
        ],
        "parameters": [
          {
            "name": "file",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "position",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogReadResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthResponse": {
        "required": [
          "success",
          "error"
        ],
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ChangePasswordRequest": {
        "required": [
          "newPassword"
        ],
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": [
              "null",
              "string"
            ]
          },
          "newPassword": {
            "type": "string"
          }
        }
      },
...
      "CreateSessionRequest": {
        "required": [
          "cols",
          "rows"
        ],
        "type": "object",
        "properties": {
          "cols": {
            "type": "integer",
            "format": "int32"
          },
          "rows": {
            "type": "integer",
            "format": "int32"
          },
          "shell": {
            "type": [
              "null",
              "string"
            ]
          },
          "workingDirectory": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "DirectoryEntry": {
        "required": [
          "name",
          "isDirectory",
          "size",
          "modified",
          "mimeType"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "isDirectory": {
            "type": "boolean"
          },
          "size": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "modified": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "mimeType": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "DirectoryListResponse": {
        "required": [
          "path",
          "entries"
        ],
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DirectoryEntry"
            }
          }
        }
      },
...
      "FileCheckRequest": {
        "required": [
          "paths"
        ],
        "type": "object",
        "properties": {
          "paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "FileCheckResponse": {
        "required": [
          "results"
        ],
        "type": "object",
        "properties": {
          "results": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/FilePathInfo"
            }
          }
        }
      },
      "FilePathInfo": {
        "required": [
          "exists",
          "size",
          "isDirectory",
          "mimeType",
          "modified",
          "isText"
        ],
        "type": "object",
        "properties": {
          "exists": {
            "type": "boolean"
          },
          "size": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "isDirectory": {
            "type": "boolean"
          },
          "mimeType": {
            "type": [
              "null",
              "string"
            ]
          },
          "modified": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "isText": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
      "FileRegisterRequest": {
        "required": [
          "sessionId",
          "paths"
        ],
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "paths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "FileResolveResponse": {
        "required": [
          "exists",
          "resolvedPath",
          "isDirectory",
          "size",
          "mimeType",
          "modified",
          "isText"
        ],
        "type": "object",
        "properties": {
          "exists": {
            "type": "boolean"
          },
          "resolvedPath": {
            "type": [
              "null",
              "string"
            ]
          },
          "isDirectory": {
            "type": "boolean"
          },
          "size": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "mimeType": {
            "type": [
              "null",
              "string"
            ]
          },
          "modified": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "isText": {
            "type": [
              "null",
              "boolean"
            ]
          }
        }
      },
      "FileUploadResponse": {
        "required": [
          "path"
        ],
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          }
        }
      },
...
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
...
      "LogEntryMessage": {
        "required": [
          "messageType",
          "source",
          "sessionId",
          "timestamp",
          "level",
          "message"
        ],
        "type": "object",
        "properties": {
          "messageType": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "sessionId": {
            "type": [
              "null",
              "string"
            ]
          },
          "timestamp": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "LogReadResponse": {
        "required": [
          "entries",
          "position",
          "fileName"
        ],
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LogEntryMessage"
            }
          },
          "position": {
            "type": "integer",
            "format": "int64"
          },
          "fileName": {
            "type": "string"
          }
        }
      },
      "LoginRequest": {
        "required": [
          "password"
        ],
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        }
      },
...
      "RenameSessionRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ResizeRequest": {
        "required": [
          "cols",
          "rows"
        ],
        "type": "object",
        "properties": {
          "cols": {
            "type": "integer",
            "format": "int32"
          },
          "rows": {
            "type": "integer",
            "format": "int32"
          },
          "viewerId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ResizeResponse": {
        "required": [
          "accepted",
          "cols",
          "rows"
        ],
        "type": "object",
        "properties": {
          "accepted": {
            "type": "boolean"
          },
          "cols": {
            "type": "integer",
            "format": "int32"
          },
          "rows": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "SessionInfoDto": {
        "required": [
          "id",
          "pid",
          "createdAt",
          "isRunning",
          "exitCode",
          "cols",
          "rows",
          "shellType",
          "name",
          "terminalTitle",
          "manuallyNamed",
          "currentDirectory",
          "foregroundPid",
          "foregroundName",
          "foregroundCommandLine",
          "order",
          "parentSessionId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "pid": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "isRunning": {
            "type": "boolean"
          },
          "exitCode": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "cols": {
            "type": "integer",
            "format": "int32"
          },
          "rows": {
            "type": "integer",
            "format": "int32"
          },
          "shellType": {
            "type": "string"
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "terminalTitle": {
            "type": [
              "null",
              "string"
            ]
          },
          "manuallyNamed": {
            "type": "boolean"
          },
          "currentDirectory": {
            "type": [
              "null",
              "string"
            ]
          },
          "foregroundPid": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "foregroundName": {
            "type": [
              "null",
              "string"
            ]
          },
          "foregroundCommandLine": {
            "type": [
              "null",
              "string"
            ]
          },
          "order": {
            "type": "integer",
            "format": "int32"
          },
          "parentSessionId": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "SessionListDto": {
        "required": [
          "sessions"
        ],
        "type": "object",
        "properties": {
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionInfoDto"
            }
          }
        }
      },
...
    }
  },
...
}